+++ src/map/mob.cpp
@@ -4760,27 +4760,27 @@ uint64 MobDatabase::parseBodyNode(const ryml::NodeRef& node) {
 	}
 
 	if (this->nodeExists(node, "AttackRange")) {
-		uint16 range;
+		uint8 range;
 
-		if (!this->asUInt16(node, "AttackRange", range))
+		if (!this->asUInt8(node, "AttackRange", range))
 			return 0;
 
 		mob->status.rhw.range = range;
 	}
 	
 	if (this->nodeExists(node, "SkillRange")) {
-		uint16 range;
+		uint8 range;
 
-		if (!this->asUInt16(node, "SkillRange", range))
+		if (!this->asUInt8(node, "SkillRange", range))
 			return 0;
 
 		mob->range2 = range;
 	}
 	
 	if (this->nodeExists(node, "ChaseRange")) {
-		uint16 range;
+		uint8 range;
 
-		if (!this->asUInt16(node, "ChaseRange", range))
+		if (!this->asUInt8(node, "ChaseRange", range))
 			return 0;
 
 		mob->range3 = range;
